requirements.txt
numpy
scipy
sympy==0.7
jupyter
PyLMI-SDP==0.2
cvxopt==1.2.6

